feat: update, add models: - add NotesCreate & NotesUpdate model schemas - update filename models -> schemas - update imports to main.py
